test(isfrr-template): add normal create/update test
