--- packages/browser/sifrr-dom/test/public/simpleelement.html
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-  </head>
-  <body>
-    <h1>OK</h1>
-    <sifrr-small></sifrr-small>
-    <script src="/sifrr.fetch.min.js" charset="utf-8"></script>
-    <script src="/sifrr.dom.js" charset="utf-8"></script>
-    <script type="text/javascript">
-      /* eslint-disable */
-      var seText = Sifrr.Dom.SimpleElement('abcd').sifrrClone();
-      var seTextState = Sifrr.Dom.SimpleElement('${this.state.text}', { text: 'yay' }).sifrrClone();
-      var seBasic = Sifrr.Dom.SimpleElement('<div><p>abcd</p></div>').sifrrClone();
-      var seMultipleChild = Sifrr.Dom.SimpleElement('<div></div><p></p>').sifrrClone();
-
-      var seStateBase = Sifrr.Dom.SimpleElement('<p>${this.state.p}</p>', { p: 'default' });
-      var seState = seStateBase.sifrrClone();
-
-      var seExtra = Sifrr.Dom.SimpleElement('<div><p>${this.state.p}</p><span></span></div>', { p: 'default' }).sifrrClone();
-      var seAttr = Sifrr.Dom.SimpleElement('<div><p data-attr=${this.state.attr} class="${this.state.class}"></p></div>').sifrrClone();
-
-      var seComplexBase = Sifrr.Dom.SimpleElement('<div><p data-attr=${this.state.attr} class="${this.state.class}" id="ok"></p>${this.state.some}</div>')
-      var seComplex = seComplexBase.sifrrClone();
-
-      var seComment = Sifrr.Dom.SimpleElement('<p>${this.state.before}<!-- comm -->${this.state.after}</p>', { before: 'b', after: 'a' }).sifrrClone();
-
-      const p = document.createElement('p');
-      p.innerHTML = '${this.state.p}';
-      var seDom = Sifrr.Dom.SimpleElement(p).sifrrClone();
-
-      var seClone = seComplexBase.sifrrClone();
-      var seStateClone = seStateBase.sifrrClone();
-      /* eslint-enable */
-    </script>
-    <script>
-      Sifrr.Dom.setup({
-        useShadowRoot: true
-      });
-      Sifrr.Dom.load('sifrr-small');
-      Sifrr.Dom.load('sifrr-row');
-    </script>
-  </body>
-</html>
