+++ packages/browser/sifrr-template/src/sifrr.template.ts
@@ -6,6 +6,8 @@ import { default as bindFor } from './binders/bindfor';
 import { default as bindForKeyed } from './binders/bindforkeyed';
 import { default as memo } from './template/memo';
 import { default as Store } from './template/store';
+import { createTemplateFromString } from './template/utils';
+import { makeChildrenEqual, makeEqual } from './template/makeequal';
 import * as types from './template/types';
 
 export { default as createUniqueString } from './ustring';
@@ -16,10 +18,14 @@ export { default as bindFor } from './binders/bindfor';
 export { default as bindForKeyed } from './binders/bindforkeyed';
 export { default as memo } from './template/memo';
 export { default as Store } from './template/store';
+export { createTemplateFromString } from './template/utils';
+export { makeChildrenEqual as makeEqualArray, makeEqual } from './template/makeequal';
 export * from './template/types';
 
 export default {
   createUniqueString,
+  createTemplateFromString,
+  makeChildrenEqual,
   html,
   css,
   update,
@@ -27,5 +33,7 @@ export default {
   bindForKeyed,
   memo,
   Store,
+  makeEqualArray: makeChildrenEqual,
+  makeEqual,
   ...types
 };
