+++ packages/browser/sifrr-template/test/browser/speedtest.arrangements.js
