+++ scripts/test/run.js
@@ -86,7 +86,6 @@ const { runTests, exec } = require('@sifrr/dev');
 const options = roots.map((root, i) => {
   const preCommand = [];
   if (!dontRunPrecommand) {
-    if (root.indexOf('sifrr-dom') < 0) preCommand.push(`cd ${root} && yarn build`);
     if (fs.existsSync(path.join(root, './test/public/package.json'))) {
       preCommand.push(`cd ${path.join(root, './test/public')} && yarn && yarn build`);
     }
@@ -107,6 +106,7 @@ const options = roots.map((root, i) => {
     folders: {
       static: [
         path.join(__dirname, '../../packages/browser/sifrr-dom/dist'),
+        path.join(__dirname, '../../packages/browser/sifrr-template/dist'),
         path.join(__dirname, '../../packages/browser/sifrr-fetch/dist')
       ],
       coverage: path.join(__dirname, '../../.nyc_output')
@@ -121,11 +121,6 @@ const options = roots.map((root, i) => {
 });
 
 async function run() {
-  if (!dontRunPrecommand) {
-    await exec(`cd ${path.join(__dirname, '../../packages/browser/sifrr-dom')} && yarn build`);
-    await exec(`yarn upgrade @sifrr/dom`);
-  }
-
   runTests(options.length === 0 ? options[0] : options, process.env.PARALLEL === 'true').then(
     ({ failures, coverage }) => {
       console.table(coverage);
